gui events now serviced in main thread
sdl gui events are now serviced in the main thread. previously, the sdl
event had its own goroutine loop.  this worked on Linux systems but
didn't seem to work on MacOS

A new interface called gui.EventsLoop has been created. When a gui that
implements this interface has been created (such as sdlplay and
sdldebug), the interface is passed back to the main thread. the
ServiceEvents() function then becomes part of the main loop.