corrected sdl event queue logic
in a recent commit 20bb515cd26366ae8b6f44fd4287c1607051a747
the intention was to loop over the sdl event queue until
exhaustion. however, we were ending the loop when an unhandled
event was encountered. this is not correct and we now end the
loop when a nil event is returned (indicating a timeout)