chip memory now writes data for future reads
some chip memory addresses are two way in the sense that whatever data
is written can be read from the same address

previously, this wasn't happening. now if a chip memory address has
a read label which is the same as its write label then the data
is written to the memory array in addition to the chip being
signalled.

this is useful for the debugger. not seen any evidence that it's
required behaviour by any ROMs, but it might be.