moved FPS limiter from PixelRenderer to TV
allows sdlimgui to work at it's own frame rate (60fps) independent of
what is happening with the emulation

the reason the FPS limiter was in PixelRenderer at all was because of
the emulator's early development. originally, the television and pixel
renderer were the same thing; then it was decided that the television
should be independent of the renderer and the renderer (wasn't called
that originally) 'embedded' the television implementation; eventually we
got the system we have now. that's my memory of it anyway, the git logs
might tell a different story.