correct fix for timer as described in URLs
https://atariage.com/forums/topic/303277-to-roll-or-not-to-roll/

https://github.com/stella-emu/stella/issues/108

code still does not handle the setting of the PA7 flag only it's
clearing when TIMINT is read. we've assumed that PA7 is set on machine
reset.  all known test cases are fine with this and there doesn't seem
to be anything in the test code that would actually set PA7, so this is
a correct interpretation (I think).