silenced go vet complaint about passing by disassembly by value
vet is complaining because the sync.RWMutex is inside and we don't want
a copy of the mutex - that would defeat the point of the mutex.

however, I think this is a misleading complaint because in the function
in question, we never access the mutex or change it or anything else in
the function.

for now changed the receiver of the function to a pointer. will seek
further clarification