Better handling of rewind slider when emulation is running
emulation is halted (temporarily) when slider is moved and will resume
again when slider is released, from the new rewind position.

stored state information is copied and assigned to the emulation, rather
than just pointing the emulation to the stored copies. the old way
worked but running the machine could then clobber what was stored in the
rewind structures.