added Coprocesser last execution window
added support for coprocessor menu

reworked window manager/menu system to be more flexible

windows that are cartridge specific now use mapper ID in title. this
means that we can store different window size/position for the cartridge
RAM window, for example, that are approprite for the mapper in question.