fixed bug caused by loading a Superhcarger ROM without a BIOS
bug was only triggered when loading was attempted when another cartridge
had already been loaded. the error path when the debugger was already
running left the debugger in an unsafe state for the disassembler

debugger will now load the ejected cartridge if an error occurs during
the loading of another cartridge.