ARM7 performance improvements
it is assumed that the program counter will only ever point to a single
memory area (the custom area in CDF/DPC+ parlance) so reading the PC
address no longer consults the memory map every instruction.

the top/level decoding of instructions happens once and the function
pointer is placed into a cache.