fixed ARM7 when PC changes program memory block
for optimisation purposes we cache the first level decoding of an
opcode. we also store the pointer to the memory mapping of the PC and
only change it if the PC falls outside the range of cache. we call this
program memory.

however, the decoding cache was not being refreshed when program memory
changed causeing visible failures in Draconian, and probably other ROMs
too.