added PauseUpdates() function to lazy system
this helps to keep lazy values when changed via the GUI to keep
synchronised with the real values

currently, this is only employed in the terminal.pushCommand() function.
if values are changed any other way, for example in the TIA window, then
the GUI will momentarily flicker as the lazy system is updated.

two solutions:

1) add PauseUpdates() in front of every change of value that is pushed
   to the emulation goroutine (via PushRawEvent())

2) only ever change values via the debugger command system.

note that both these solutions imply that the 'play' mode needs a
PushRawEvent() implemented. currently, it doesn't have that.