CRT preview in debugger TV screen can never shown uncropped/debugging colors or overlay
it was felt that allowing CRT to be combined with debugging guides adds
nothing. however, simply removing the CRT preview option didn't feel
correct either.

in addition, the opportunity to clarify how window scaling works was
taken. in particular the scaledWidth() function and similar functions
were removed in favour of single values which are updated as necessary
every render frame.