reverted to the old resizing method
the simpler method was fine except that it didn't work well with the
strict scaling policy that we introduced in d45e609534ad7321d2f7309ab04e2646fe18f31b.

it was particularly noticeable when moving to full screen when a fully
"extended" NTSC image would not look right unless the ROM really did use
the entirity of the screen. for example Spike's Peak looks okay but
otherwise, there would be enormous black bars.