fixed deadlock in disassembly package
caused by not releasing crital section lock on error. error caused by
the presence of the "ejected" cartridge. most probably caused by trying
to load a unrecognised file.

in a headless disassembly, this doesn't matter because no other
goroutine ever tries to acquire the lock but in the sdlimgui debugger,
the disassembly window tries to acquire the lock during iteration.