playscreen will update texture even if screen is unstable
the stability check is simply wrong now that I think about it more, but
it was only obvious for a ROM that never achieved stability (and
therefore never showing any output).

in these cases the texture was never updated to begin with so the screen
never had anything to show.

additional observation: it follows from this reasoning that on the
occasions when a television frame lost stability it would show "stale"
information for the period the TV was unsynced.