television sends every signal required for frame even for "short frames"
sometimes a television frame will be "shorter" than previous frames.
previously the television only sent the signals for the shorter frame,
this could play havoc with PixelRenderers under certain circumstances;
particularly visible under rewind conditions.

this was noticed when rewinding a moviecart over frames between the
attract screen and the beginning of the movie.

to fix this. the television now sends null signals (video black) for
every required signal in the current frame (whatever size that might be)

there doesn't seem to be any performance penalty