removed RawEventImm channel. removed PushGotoCoords()
RawEventImm was added originally to allow GUI events to preempt all
other pushed RawEvents. this was needed because of how lazy value update
requests were pushed. we've throttle these requests now so we no longer
need the preempt channel

PushGotoCoords() was added because we wanted to use the RawEventImm
channel. pusing the GOTO command is now fast enough for discrete calls.
we may want to reimplement it if we need to send many GOTO commands -
similar to how PushRewind() is required in place of REWIND commands