removed visible artefact on debug screen when pausing emulation
caused by changes introduced in dda5ed5d5c4ce93fe8ea4fc40e3ce763eb0b10cf

taking the clock/scanline values from LazyTV was a good idea but the
time it takes for the lazy system to update is just long enough for the
old value to be visible.

an additional argument to PixelRender.SetPixels() is required but it
can't be helped.