step back instruction always produces correct results
stepping back previously was "correct" but was confusing when looking at
the disassembly (rather than the debugging screen)

introduced the concept of a catchupQuantum. distinct from the existing
stepQuantum (renamed from 'quantum') but having the same underlying type

the catchupQuantum makes sure that the inputloop is left in the correct
state. in other words in an 'isVideoStep' version so of inputLoop(), or
not depending on how the catchup was initiated