coprocessor disassembly is disabled by default
renamed "last execution" window to "coprocessor disassembly" where
the actual name of the coprocessor is used

coproc disassembly window now shows an enable and a disable button. if
the emulation is paused the previous two frames of the emulation will be
rerun in order to generate a disassembly

coprocessor disassembly and last execution are shown in tabs. if
disassembly is disabled then the last execution tab will show nothing
but the disassembly tab will show what we have, with a note that it is
not bing udpated