small audio error noticeable during long periods of "non-zero" silence
(zero being whatever SDL detects as the silence value for the sound
card - usually zero).

the error was noticeable when a ROM used a non-zero silence value and
simply an unchanging value instead (which the speaker treats as silence
- because the cone never moves).

queuing audio signals from the television signal array used all signals
from the beginning of the array but because of the way the television
writes to the array the first couple of entries have probably never been
written to. in other words, a audio value of zero was sneaking it's way
into the audio queue, creating a buzz.

affected moviecart ROMS