ROM select window correctly handles collapsed window
the window would slowly grow in size when collapsed/uncollapsed. it's
probably something to do with the height measuring but whatever the
reason the correct thing to do is to check the result of imgui.Begin()
(which solves the problem)

a similar change should probably by applied to all windows.