made audio tracker available in playmode (F10)
this means that playmode now uses the lazy system. it's only loading
what it needs so as not to use too much memory. the good news is that it
doesn't seem to take any significant CPU time

tracker reset on debugger/emulation reset

piano keys window is now part of the tracker window

updated README indicating new hotkey