dbgmem.Poke works with "read" addresses and calls MapAddress() appropriately
although the words "read" and "write" might lead us to think that we
"peek" from "read" addresses and "poke" to "write" addresses, it is in
fact necessary to "poke" to "read" addresses.

on the surface this doesn't appear to be correct but on further thought
it is obviously true - we are in fact changing the value that is
subsequently read by the CPU, so that means poking to a read address.