DWARF - make sure the correct LineReader is used when dereferencing DeclFile values
the previous fix (809083adeb) was unecessary as the DeclFile value
should never be out of range of the Files() array. if it is then
something has gone seriously wrong and it is acceptable to crash IMO

the new method, looks up the compile unit for a AbstractOrigin entry in
an index for each compile unit. only then can we use ask for the Files()
array from the correct compile unit