cartridges are now attached to the television FrameTrigger system
this means that cartridge types that need that information can implement
the FrameTrigger interface and act accordingly.

this was added for the ACE format so that it can update the ARM
preferences less frequently. this gives a small but significant
performance improvement. it also applies to DPC+ and the CDF formats
although performence improvement is less noticeable in those instances
because such ROMs tend not to call the ARM program that frequently