greatly simplified how coproc developer package profiles execution
cartridge mappers now longer have to deal with the developer package

coprocessor implementations now only need to ask for a profiling map
(previously called an execution map) which can be used to accumulate
cycles

profiling data is processed every scanline and is triggered by the TV.
this method of processing (as opposed to relying on the coprocessor or
mapper to trigger it) means the profiling map can be more gently reset
as part of that process and with little performance impact. the previous
method necessitated the map be reallocated if performance was to be
upheld. the reallocation would inevitably put additional strain on the
garbage collector although this wasn't measured