fixed execution profiling for CDFJ and DPC+
commit 8f09a6521 made assumptions that aren't true for these formats.
the method now is for the cartridge mapper to trigger the accumulation
when execution has ended

the consequence of this is that the kernel the execution is considered
to have executed in is the kernel the execution ends in. this is good
enough for most purposes

this temporarily breaks profiling for the ELF and ACE formats. it maybe
necessary to have two different profiling triggers: one for CDFJ/DPC+
style mappers and one for StrongARM style mappers