tentative support for 32bit thumb instructions in coproc/developer package
disassmbly of thumb2 is not complete yet but the SourceDisasm now notes
whether an opcode is 32bit or not. in particulary, 32bit instructions no
longer have misleading static disassembly entries

the live disassembly window also correctly differentiates 16 and 32 bit
instructions

the above changes are not final. much better handling of the disassembly
in both the developer and arm packages is required

to complement the above (temporary) changes, the source and performance
windows now show the instruction bytecode, 16bit or 32bit, in the
tooltips