thumb instruction BLX (register) corrected (ARMv7-M)
previously, this instruction had only been seen in use when jumping to a
StrongARM function. because of a mistake made when building the table
for those functions the BLX instruction was incorrectly setting the ARM
program counter (PC). when the instruction was seen in another context
the PC was incorrectly set

correcting the StrongARM function table allows the BLX instruction to
set the PC correctly