moved ARM timers and RNG to arm.peripherals package
memorymodel expanded to contain addresses for new peripherals. also, the
presence or absence of a peripheral can be indicated

MAM is a peripheral but it's still in the arm package because it's so
tightly bound with cycle counting. however, the memorymodel can indicate
that the MAM is not present, meaning that memory subsystem won't try to
address it

memorymodel can specify a clock divisor value that will be applied to
the main clock when ticking the timers