CartYieldHook.CartYield() returns bool
the returned value is true if the ARM emulation should 'give up' for the
time being. when this happens it is assumed that the state of the
emulation is not ideal and so drifting by a couple of video cycles (at
most) is not an issue - we'll try to remove even this small deviation
in the future but it's not a priority

A YieldReason, other than YieldForVCS, when in playmode will cause the
emulation to switch to debugging mdoe. there are possible issues with
this but for now it works