deadlock in IllegalAccess if the coproc PC has no source line
the derefencing of the nil pointer cause by the missing source line
causes a panic. however, the dereferencing was inside a critical section
and the mutex was never release as a result of the panic

mutex unlocking is now a deferred function and a nil check on the source
line to prevent a panic

in addition, debugger.CartYield() now checks the running status of the
emulation. if it is true then the function returns true, instructing the
coprocessor loop in the cartridge to exit immediately