thumb format 5 MOV instruction did not function PC correctly
the adjustment to the PC was only done if the source register was the LR
register. this is incorrect and the adjustment is required regardless of
the source

I've also applied the PC adjustment for the format 5 ADD instruction
which according to section 5.5.5 of the ARM7TDMI reference it should.
however, I've not applied the adjustment to the CMP instruction -
adjusting the value in that context doesn't seem right to me

the BX/BLX instruction was already adjusting the PC correctly

I don't believe there are any other 16 bit thumb instructions that can
affect the PC. the only others are the branch instructions and the POP
instructions, which seem to be correct. the LDMIA instruction I'm not
sure about and will need more research