changes in 6737d50c8a broke zookeeper. fixed
6737d50c8a introduced the ability to specify the null read value, which
seems to be required for Uno/Pluscart ROMs to function correctly in some
circumstances

however, the same change also introduced an early return (with the
specified null read value) if the ARM program tried to access below a
certain address. this was incorrect

also removed the early return if a memory access address was misaligned
(for 8bit/16bit reads). I've never seen a misaligned address but
returning early without even attempting to access the memory seems wrong