fixed condition under which a local variable is said to be resolvable
this undoes some of the work done in eae525bf51. in that commit we said
that a variable was resolvable if a loclist existed for any address
associated with a source line. however this doesn't work because the
variable may not exist until the third instruction (for example) of the
source line

the downside of this is that in an optimised program a variable may
appear to go out of scope when you wouldn't expect it to, but I don't
see a way around that currently