corrected live ARM disassembly for Thumb-2 instructions
uses placeholder information for Thumb-2 instructions for now. still
need to figure out an efficient way to add disasm information to a
decoded instruction. the current method for original Thumb instructions
is fine but it seems needlessly complex to use that for the larger
instruction set

cycle order is not reset if a 32bit instruction is being decoded. this
didn't affect cycle accumulation (which is correct as far as possible)
but it did affect cycle reporting for 32bit instructions

(cycle counting for the Thumb-2 instructions is still incomplete.
however, the majority of cycles will be accounted for because they occur
as the instruction is being deocded)

live disassembly and source level disassembly (coprocessor/developer
package) structures are still distinct. they use the same decoding
functions so the information should be identicaly but how we handle the
information is different depending on context