ARM disassembly/decoding less verbose
some time ago, thumb disassembly was reworked to be entirely separate
from the thumb instruction decoding. the aim was to increase the
decoding speed during execution. this was achieved that but with
considerable added complexity and duplication of branching logic

if the same method was used for Thumb2 then the package would become
horribly complex

with this commit the disassembly is again part of the main decoding
logic

the changes seems to work well with no significant performance impact