improved static disassembly of ARM code
ARM now uses binary.ByteOrder() for reading data. All extant examples or
2600/ARM ROMs were handled correctly but ByteOrder() is neater and
better supports the possibility of the ELF format

a previous commit experimented with passing the ARM instance to the
disasm/decode functions as a parameter, thereby allowing to use the nil
value in some way. this was a poor experiment and the code has reverted
to the decode functions being directly connected to the ARM type. than
accepting the ARM as a parameter. a sparse ARM instance can be created
for the purposes of static disassembly

'correct' disassembly of Thumb2 instruction set. operand details are not
yet complete but the operator and a basic description in place of the
actual operand are in place. this means that the 'fudge' disassembly
method can be removed

fixed Thumb2 version of BL/BLX instructions

an entire class of instructions was handled but with no operation or a
panic. this would cause some programs to run but incorrectly

ARM7TDMI handles disassembly of 32bit instructions in the same way as
ARM7_M. in practice this applies to BL instruction only but none-the-less
the new presentation method is neater