better control of timeline thumbnail generation
previously, throttling of thumbnail generation was controlled by the
clock. for fast running emulation this wasn't a problem but for slower
emulation (eg. ACE) this was ineffective and caused multiple thumbnail
request to queue up

this revealed a general inefficiency, even if it wasn't noticeable in
"normal" circumstances

the new throttling method prevents more a thumbnail request to be
generated if one is in process