fixed crash when debugging 8k mnetwork ROMS
GetBank() implemention for mnetwork reported bank 7 as being the bank
ocuppying segment 1 (the second segment). this is fine in most instances
except that in 8k mnetwork ROMs there is no bank 7. the correct value
for segment 1 is the number of the last bank number available in the
ROM, which in the case of 8k ROMs is bank 3

8k mnetwork ROMs are not common. I only know of Bump'n'Jump