removed "ROM Setup" from coprocessor profiling
"ROM Setup" is too confusing a state to be useful. Initialisation, based
on the current emulation state, might be more helpful but it fails for
the same reason as "ROM Setup" (which was based on screen stability) in
that it doesn't really fit in with the current method of keeping
profiling statistics

renamed KernelVCS (from profiling package) to Focus. this better
describes what the intention of the type is

renamed Performance window to Profiling