improved DPC static memory implementation
previously, the static memory was not subject to the Snapshot()
function required by the rewind system. static memory is now copied
on every snapshot. this isn't normally needed because in DPC the data
cannot be changed by the 6507 progam. however we might want to change
the data through the debugger and it is therefore desirable that copies
are kept for comparison and rewind purposes