simplified framebuffer pipeline
clarified how and when the vertex shader should flip the texture

occassional graphical artefacts no longer appear. the artefacts appeared
(or were more noticeable) when the TV image was scaled to certain sizes.
they were caused by incorrect use of the framebuffer. using the same
texture as the source and destination of the shading process does not
work as expected