moved monitorSyncHigher test in screen.go
the test now affects whether the wait channel is used. the frameQueue is
not affected directly

this fixes problems with the screen not appearing to update (in
playmode) when the TV is of a higher refresh rate than the monitor. this
happened intermittently depending on where the frame queue indexes where
at the time the TV settled on the problem frequency, so it was difficult
to debug. however it happened more-or-less consistently when running the
Todd Frye version of xevious