bumped imgui-go requirement to 4.7.2
this is purely house-keeping because I fumbled the release of
imgui-go 4.7.1. it's okay except the tag to 4.7.1 is not in a branch,
which may seem odd.

it was caused by starting the release process on github but then force
pushing an additional change. the release process is now pointing at
a commit that is not part of a branch