correct fix for rewinding of fastload supercharger files
the problem was that the Plumb() function was not providing the new
pointer to the cartridge

this wasn't a problem for the soundloud sub-implementation because that
never changes supercharger memory directly

however, for fastload, the load process does poke supercharger memory as
part of the fast load process. this meant that the fastload
implementation could poke memory of another rewind state, meaning that
the current rewind state would never get past the load() point. ie. the
PC would never be set to the correct value etc.

fastload does not send NotifySuperchargerFastload notification unless it
is in the MainEulation (not during thumbnailing etc.)

previous attempt at fix in commit 78a6880