LateCOLUPF expanded to include COLUBK register
when I implemented the feature I only knew of Quickstep as
being a problem ROM for some RGB mods. however, there is now
an example of a ROM where the COLUBK register is a problem.

rather than add a new option I simply renamed LateCOLUPF to
LateColor and set it to affect the COLUBK in addition to
COLUPF