disabled sdl.HINT_VIDEO_X11_NET_WM_BYPASS_COMPOSITOR hint
this is on by default and seems to be the cause of the occassional loss
of VSYNC (for want of a better description) seen when displaying ROMs
with flicker kernels

note that disabling the bypass_compositor option doesn't completely
solve the VSYNC problems when running under the muffin window manager.
it does however solve the problem under other window managers that have
been tested. either way, the issue is purely one for X11 and doesn't
affect Windows of MacOS

I've also removed the artificial window event delays added in 28c22cb34.
the delays seemed to mitigate the problem but not solve it.